Add custom-valued enums & generate array type specs

+++ lib/swablib.h
@@ -5,7 +5,8 @@
 
 #define BINDLIB_VERSION "0.0.0"
 
-typedef int bool;
+// Not sure what the "correct" option is here, but this is best for bool[]
+typedef int8_t bool;
 
 typedef uint8_t u8;
 typedef  int8_t i8;
